Fix phpstan errors
